DELdata: An R package to download Development Engagement Lab datasets

The main function of this package is get_del_data(), which allows you to download DEL's data based on the country, type, and years.

get_del_data(country = NULL, year = NULL, type = NULL)

Arguments

country

A vector of countries. Options include "France", "Germany", "Great Britain", and "United States".

year

A vector containing one or several years. Acceptable values are from 2019 to 2024.

type

A vector containing the data type. Options include "Panel", "Tracker", and "Sandbox".

Value

A list of data frames containing the requested datasets.

Examples

get_del_data(country = 'France', year = '2020', type = 'Panel')
#> Processing DOI: 10.7910/DVN/KWTWCF
#> Error occurred while fetching DOI: 10.7910/DVN/KWTWCF
#> lexical error: invalid char in json text.
#>                                        <!DOCTYPE HTML PUBLIC "-//IETF/
#>                      (right here) ------^
#> list()